Revert "Internal: Infra for the ability to define page transitions [ED-3849] (#15427)"
This reverts commit 7a2f736448ff100b10bb8f0d5fb93c0d5b93cfb5.
